Move mapToJson.h definitions outside of main.cpp
