+++ JsonMap.vcxproj
@@ -128,6 +128,7 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="src\main.cpp" />
+    <ClCompile Include="src\mapToJson.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\mapCollect.h" />
