+++ JsonMap.vcxproj.filters
@@ -2,6 +2,7 @@
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
     <ClCompile Include="src\main.cpp" />
+    <ClCompile Include="src\mapToJson.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\mapCollect.h" />
